iouring: use NOP as cancel op
If a job and its respective cancel job are in the same submission bulk,
they cancel out when entering the kernel, leading to no submission.
Because of that, the iouring done handler is not called for this req,
leading to an uncalled callback.
To prevent this, the iouring's nop is used to indicate a cancellation.
If the result of the req is set to UV_ECANCELED, the handler knows that
only the NOP must be handled. Potential others (e.g. successful work,
unsuccessful cancellation, ...) are ignored in this case.
Since the CQ entry does not contain the opcode, we have to create a
respective user_data container to identify the opcode in the done handler.
The user container of the actual work job is stored in the req to be
able to identify it.